refactor: PlayerState 생성 규칙 주체 변경(Result -> PlayerState)
